Extract WebApiContrib.IoC.* solutions.
