+++ WebApiContrib.IoC.Autofac.sln
@@ -1,0 +1,35 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2012
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{4C93B8FC-96F9-47BA-9380-753847F03E1C}"
+	ProjectSection(SolutionItems) = preProject
+		.nuget\NuGet.exe = .nuget\NuGet.exe
+		.nuget\NuGet.targets = .nuget\NuGet.targets
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApiContrib.IoC.Autofac", "src\WebApiContrib.IoC.AutoFac\WebApiContrib.IoC.Autofac.csproj", "{AA772346-54BD-48C4-AEC0-CCF54A7E45FF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApiContrib.IoC.Autofac.Tests", "test\WebApiContrib.IoC.Autofac.Tests\WebApiContrib.IoC.Autofac.Tests.csproj", "{960EDBC5-0C46-4CD7-B0B6-B4D5F30B9092}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{960EDBC5-0C46-4CD7-B0B6-B4D5F30B9092}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{960EDBC5-0C46-4CD7-B0B6-B4D5F30B9092}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{960EDBC5-0C46-4CD7-B0B6-B4D5F30B9092}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{960EDBC5-0C46-4CD7-B0B6-B4D5F30B9092}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AA772346-54BD-48C4-AEC0-CCF54A7E45FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AA772346-54BD-48C4-AEC0-CCF54A7E45FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AA772346-54BD-48C4-AEC0-CCF54A7E45FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AA772346-54BD-48C4-AEC0-CCF54A7E45FF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		EnterpriseLibraryConfigurationToolBinariesPath = packages\Unity.2.1.505.0\lib\NET35
+	EndGlobalSection
+EndGlobal
