+++ test/WebApiContrib.IoC.Autofac.Tests/Helpers/IContactRepository.cs
@@ -1,0 +1,6 @@
+namespace WebApiContrib.IoC.Autofac.Tests.Helpers
+{
+    public interface IContactRepository
+    {
+    }
+}
