+++ test/WebApiContrib.IoC.CastleWindsor.Tests/Helpers/IContactRepository.cs
@@ -1,0 +1,6 @@
+namespace WebApiContrib.IoC.CastleWindsor.Tests.Helpers
+{
+    public interface IContactRepository
+    {
+    }
+}
