+++ test/WebApiContrib.IoC.CastleWindsor.Tests/WebApiContrib.IoC.CastleWindsor.Tests.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{960EDBC5-0C46-4CD7-B0B6-B4D5F30B9092}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>WebApiContribTests</RootNamespace>
-    <AssemblyName>WebApiContribTests</AssemblyName>
+    <RootNamespace>WebApiContrib.IoC.CastleWindsor.Tests</RootNamespace>
+    <AssemblyName>WebApiContrib.IoC.CastleWindsor.Tests</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <RestorePackages>true</RestorePackages>
@@ -33,6 +33,16 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Castle.Core, Version=3.0.0.0, Culture=neutral, PublicKeyToken=407dd0808d44fbdc, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\Castle.Core.3.0.0.4001\lib\net40-client\Castle.Core.dll</HintPath>
+    </Reference>
+    <Reference Include="Castle.Windsor, Version=3.0.0.0, Culture=neutral, PublicKeyToken=407dd0808d44fbdc, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\Castle.Windsor.3.0.0.4001\lib\net40\Castle.Windsor.dll</HintPath>
+    </Reference>
+    <Reference Include="Microsoft.Web.Infrastructure, Version=1.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
+      <Private>True</Private>
+      <HintPath>..\..\packages\Microsoft.Web.Infrastructure.1.0.0.0\lib\net40\Microsoft.Web.Infrastructure.dll</HintPath>
+    </Reference>
     <Reference Include="Newtonsoft.Json">
       <HintPath>..\..\packages\Newtonsoft.Json.4.5.6\lib\net40\Newtonsoft.Json.dll</HintPath>
     </Reference>
@@ -60,60 +70,26 @@
     <Reference Include="System.Web.Http, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\..\packages\Microsoft.AspNet.WebApi.Core.4.0.20505.0\lib\net40\System.Web.Http.dll</HintPath>
     </Reference>
-    <Reference Include="System.Web.Http.SelfHost, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\Microsoft.AspNet.WebApi.SelfHost.4.0.20505.0\lib\net40\System.Web.Http.SelfHost.dll</HintPath>
-    </Reference>
-    <Reference Include="System.Web.Http.WebHost, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\Microsoft.AspNet.WebApi.WebHost.4.0.20505.0\lib\net40\System.Web.Http.WebHost.dll</HintPath>
-    </Reference>
     <Reference Include="System.Web.Routing" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Formatting\PlainTextFormatterTests.cs" />
     <Compile Include="Helpers\Contact.cs" />
     <Compile Include="Helpers\ContactsController.cs" />
     <Compile Include="Helpers\IContactRepository.cs" />
     <Compile Include="Helpers\InMemoryContactRepository.cs" />
     <Compile Include="Helpers\PrecannedMessageHandler.cs" />
-    <Compile Include="MessageHandlers\EncodingHandlerTests.cs" />
-    <Compile Include="MessageHandlers\ETagHandlerTests.cs" />
-    <Compile Include="MessageHandlers\HeadMessageHandlerTests.cs" />
-    <Compile Include="MessageHandlers\LoggingHandlerTests.cs" />
-    <Compile Include="MessageHandlers\MessageHandlerTester.cs" />
-    <Compile Include="MessageHandlers\NotAcceptableMessageHandlerTests.cs" />
-    <Compile Include="ResponseMessages\ConflictResponseMessageTests.cs" />
-    <Compile Include="ResponseMessages\CreateResponseMessageTests.cs" />
-    <Compile Include="ResponseMessages\HttpResponseMessageTester.cs" />
-    <Compile Include="ResponseMessages\NotModifiedResponseTests.cs" />
-    <Compile Include="ResponseMessages\OkResponseMessageTests.cs" />
-    <Compile Include="ResponseMessages\MovedPermanentlyResponseMessageTests.cs" />
-    <Compile Include="ResponseMessages\RedirectResponseTests.cs" />
-    <Compile Include="ResponseMessages\TemporaryRedirectResponseMessageTests.cs" />
-    <Compile Include="Routing\HttpRouteTableBuilderTests.cs" />
-    <Compile Include="Routing\Mockups\FakeController.cs" />
-    <Compile Include="TestBase.cs" />
-    <Compile Include="Conneg\ContentNegotiationTests.cs" />
-    <Compile Include="TestHelpers.cs" />
+    <Compile Include="DependencyInjectionTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
-  <ItemGroup />
   <ItemGroup>
-    <ProjectReference Include="..\..\src\WebApiContrib.Testing\WebApiContrib.Testing.csproj">
-      <Project>{DDEB5F68-7E22-4F54-9B04-63518083B83E}</Project>
-      <Name>WebApiContrib.Testing</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\src\WebApiContrib.WebHost\WebApiContrib.WebHost.csproj">
-      <Project>{2544C950-013F-4287-B5F4-8D1D0AEC44E2}</Project>
-      <Name>WebApiContrib.WebHost</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\src\WebApiContrib\WebApiContrib.csproj">
-      <Project>{C8EBEB5F-1664-4B36-9FA0-9979830EC5B8}</Project>
-      <Name>WebApiContrib</Name>
+    <ProjectReference Include="..\..\src\WebApiContrib.IoC.CastleWindsor\WebApiContrib.IoC.CastleWindsor.csproj">
+      <Project>{3735bc1c-6224-417b-8816-d6bc6498682f}</Project>
+      <Name>WebApiContrib.IoC.CastleWindsor</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
