+++ test/WebApiContrib.IoC.CastleWindsor.Tests/packages.config
@@ -1,0 +1,16 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="AspNetWebApi.Core" version="4.0.20505.0" />
+  <package id="Castle.Core" version="3.0.0.4001" />
+  <package id="Castle.Windsor" version="3.0.0.4001" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="4.0.20505.0" />
+  <package id="Microsoft.AspNet.WebApi.Core" version="4.0.20505.0" />
+  <package id="Microsoft.Net.Http" version="2.0.20505.0" />
+  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" />
+  <package id="Newtonsoft.Json" version="4.5.6" />
+  <package id="NUnit" version="2.6.0.12054" />
+  <package id="NUnit.Runners" version="2.6.0.12051" />
+  <package id="RhinoMocks" version="3.6.1" />
+  <package id="Should" version="1.1.12.0" />
+  <package id="System.Net.Http.Formatting" version="4.0.20505.0" />
+</packages>
