+++ test/WebApiContrib.IoC.Ninject.Tests/Helpers/IContactRepository.cs
@@ -1,0 +1,6 @@
+namespace WebApiContrib.IoC.Ninject.Tests.Helpers
+{
+    public interface IContactRepository
+    {
+    }
+}
