+++ test/WebApiContrib.IoC.StructureMap.Tests/Helpers/IContactRepository.cs
@@ -1,0 +1,6 @@
+namespace WebApiContrib.IoC.StructureMap.Tests.Helpers
+{
+    public interface IContactRepository
+    {
+    }
+}
