+++ test/WebApiContrib.IoC.Unity.Tests/Helpers/Contact.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace WebApiContrib.IoC.Unity.Tests.Helpers
+{
+    public class Contact
+    {
+        public DateTime Birthday { get; set; }
+
+        public int Id { get; set; }
+    }
+}
