+++ test/WebApiContrib.IoC.Unity.Tests/Helpers/IContactRepository.cs
@@ -1,0 +1,6 @@
+namespace WebApiContrib.IoC.Unity.Tests.Helpers
+{
+    public interface IContactRepository
+    {
+    }
+}
