+++ test/WebApiContrib.IoC.Unity.Tests/Helpers/InMemoryContactRepository.cs
@@ -1,0 +1,10 @@
+namespace WebApiContrib.IoC.Unity.Tests.Helpers
+{
+    public class FileContactRepository : IContactRepository
+    {
+    }
+
+    public class InMemoryContactRepository : IContactRepository
+    {
+    }
+}
