+++ test/WebApiContrib.IoC.Unity.Tests/WebApiContrib.IoC.Unity.Tests.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{960EDBC5-0C46-4CD7-B0B6-B4D5F30B9092}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>WebApiContribTests</RootNamespace>
-    <AssemblyName>WebApiContribTests</AssemblyName>
+    <RootNamespace>WebApiContrib.IoC.Unity.Tests</RootNamespace>
+    <AssemblyName>WebApiContrib.IoC.Unity.Tests</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <RestorePackages>true</RestorePackages>
@@ -33,6 +33,19 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Microsoft.Practices.ServiceLocation">
+      <HintPath>..\..\packages\CommonServiceLocator.1.0\lib\NET35\Microsoft.Practices.ServiceLocation.dll</HintPath>
+    </Reference>
+    <Reference Include="Microsoft.Practices.Unity, Version=2.1.505.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\Unity.2.1.505.0\lib\NET35\Microsoft.Practices.Unity.dll</HintPath>
+    </Reference>
+    <Reference Include="Microsoft.Practices.Unity.Configuration">
+      <HintPath>..\..\packages\Unity.2.1.505.0\lib\NET35\Microsoft.Practices.Unity.Configuration.dll</HintPath>
+    </Reference>
+    <Reference Include="Microsoft.Web.Infrastructure, Version=1.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
+      <Private>True</Private>
+      <HintPath>..\..\packages\Microsoft.Web.Infrastructure.1.0.0.0\lib\net40\Microsoft.Web.Infrastructure.dll</HintPath>
+    </Reference>
     <Reference Include="Newtonsoft.Json">
       <HintPath>..\..\packages\Newtonsoft.Json.4.5.6\lib\net40\Newtonsoft.Json.dll</HintPath>
     </Reference>
@@ -60,60 +73,26 @@
     <Reference Include="System.Web.Http, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\..\packages\Microsoft.AspNet.WebApi.Core.4.0.20505.0\lib\net40\System.Web.Http.dll</HintPath>
     </Reference>
-    <Reference Include="System.Web.Http.SelfHost, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\Microsoft.AspNet.WebApi.SelfHost.4.0.20505.0\lib\net40\System.Web.Http.SelfHost.dll</HintPath>
-    </Reference>
-    <Reference Include="System.Web.Http.WebHost, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\Microsoft.AspNet.WebApi.WebHost.4.0.20505.0\lib\net40\System.Web.Http.WebHost.dll</HintPath>
-    </Reference>
     <Reference Include="System.Web.Routing" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Formatting\PlainTextFormatterTests.cs" />
     <Compile Include="Helpers\Contact.cs" />
     <Compile Include="Helpers\ContactsController.cs" />
     <Compile Include="Helpers\IContactRepository.cs" />
     <Compile Include="Helpers\InMemoryContactRepository.cs" />
     <Compile Include="Helpers\PrecannedMessageHandler.cs" />
-    <Compile Include="MessageHandlers\EncodingHandlerTests.cs" />
-    <Compile Include="MessageHandlers\ETagHandlerTests.cs" />
-    <Compile Include="MessageHandlers\HeadMessageHandlerTests.cs" />
-    <Compile Include="MessageHandlers\LoggingHandlerTests.cs" />
-    <Compile Include="MessageHandlers\MessageHandlerTester.cs" />
-    <Compile Include="MessageHandlers\NotAcceptableMessageHandlerTests.cs" />
-    <Compile Include="ResponseMessages\ConflictResponseMessageTests.cs" />
-    <Compile Include="ResponseMessages\CreateResponseMessageTests.cs" />
-    <Compile Include="ResponseMessages\HttpResponseMessageTester.cs" />
-    <Compile Include="ResponseMessages\NotModifiedResponseTests.cs" />
-    <Compile Include="ResponseMessages\OkResponseMessageTests.cs" />
-    <Compile Include="ResponseMessages\MovedPermanentlyResponseMessageTests.cs" />
-    <Compile Include="ResponseMessages\RedirectResponseTests.cs" />
-    <Compile Include="ResponseMessages\TemporaryRedirectResponseMessageTests.cs" />
-    <Compile Include="Routing\HttpRouteTableBuilderTests.cs" />
-    <Compile Include="Routing\Mockups\FakeController.cs" />
-    <Compile Include="TestBase.cs" />
-    <Compile Include="Conneg\ContentNegotiationTests.cs" />
-    <Compile Include="TestHelpers.cs" />
+    <Compile Include="DependencyInjectionTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
-  <ItemGroup />
   <ItemGroup>
-    <ProjectReference Include="..\..\src\WebApiContrib.Testing\WebApiContrib.Testing.csproj">
-      <Project>{DDEB5F68-7E22-4F54-9B04-63518083B83E}</Project>
-      <Name>WebApiContrib.Testing</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\src\WebApiContrib.WebHost\WebApiContrib.WebHost.csproj">
-      <Project>{2544C950-013F-4287-B5F4-8D1D0AEC44E2}</Project>
-      <Name>WebApiContrib.WebHost</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\src\WebApiContrib\WebApiContrib.csproj">
-      <Project>{C8EBEB5F-1664-4B36-9FA0-9979830EC5B8}</Project>
-      <Name>WebApiContrib</Name>
+    <ProjectReference Include="..\..\src\WebApiContrib.IoC.Unity\WebApiContrib.IoC.Unity.csproj">
+      <Project>{543dd4bc-c610-4381-82d9-2ba3686aee0f}</Project>
+      <Name>WebApiContrib.IoC.Unity</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
